Backto bug fixes: namepsace error on base class, merging bug wtih jsonpayloads
